[SELC-4041] refactor: send contracts using ContractNotificationService (#390)
* contract service notification

* object mapper

* fix conditional property

* remove unused code
