+++ app/src/main/resources/config/application.yml
@@ -47,3 +47,5 @@ logging:
 core:
   user-event-service:
     type: ${CORE_USER_EVENT_SERVICE_TYPE:ignore}
+  contract-event-service:
+    type: ${CORE_CONTRACT_EVENT_SERVICE_TYPE:ignore}
