+++ core/src/main/java/it/pagopa/selfcare/mscore/core/ContractService.java
@@ -1,7 +1,6 @@
 package it.pagopa.selfcare.mscore.core;
 
 import com.fasterxml.jackson.core.JsonGenerator;
-import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonSerializer;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.SerializerProvider;
@@ -10,28 +9,17 @@ import com.openhtmltopdf.pdfboxout.PdfRendererBuilder;
 import com.openhtmltopdf.svgsupport.BatikSVGDrawer;
 import eu.europa.esig.dss.validation.SignedDocumentValidator;
 import eu.europa.esig.dss.validation.reports.Reports;
-import it.pagopa.selfcare.commons.base.logging.LogUtils;
 import it.pagopa.selfcare.commons.base.utils.InstitutionType;
 import it.pagopa.selfcare.commons.utils.crypto.model.SignatureInformation;
 import it.pagopa.selfcare.commons.utils.crypto.service.PadesSignService;
 import it.pagopa.selfcare.commons.utils.crypto.service.PadesSignServiceImpl;
 import it.pagopa.selfcare.commons.utils.crypto.service.Pkcs7HashSignService;
 import it.pagopa.selfcare.mscore.api.FileStorageConnector;
-import it.pagopa.selfcare.mscore.api.InstitutionConnector;
-import it.pagopa.selfcare.mscore.api.PartyRegistryProxyConnector;
 import it.pagopa.selfcare.mscore.config.CoreConfig;
 import it.pagopa.selfcare.mscore.config.PagoPaSignatureConfig;
-import it.pagopa.selfcare.mscore.constant.RelationshipState;
-import it.pagopa.selfcare.mscore.core.config.KafkaPropertiesConfig;
-import it.pagopa.selfcare.mscore.core.util.InstitutionPaSubunitType;
 import it.pagopa.selfcare.mscore.exception.InvalidRequestException;
-import it.pagopa.selfcare.mscore.exception.MsCoreException;
-import it.pagopa.selfcare.mscore.exception.ResourceNotFoundException;
-import it.pagopa.selfcare.mscore.model.InstitutionToNotify;
-import it.pagopa.selfcare.mscore.model.NotificationToSend;
-import it.pagopa.selfcare.mscore.model.QueueEvent;
-import it.pagopa.selfcare.mscore.model.RootParent;
-import it.pagopa.selfcare.mscore.model.institution.*;
+import it.pagopa.selfcare.mscore.model.institution.Institution;
+import it.pagopa.selfcare.mscore.model.institution.InstitutionGeographicTaxonomies;
 import it.pagopa.selfcare.mscore.model.onboarding.OnboardingRequest;
 import it.pagopa.selfcare.mscore.model.onboarding.ResourceResponse;
 import it.pagopa.selfcare.mscore.model.onboarding.Token;
@@ -40,12 +28,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.text.StringSubstitutor;
 import org.jsoup.Jsoup;
 import org.jsoup.helper.W3CDom;
-import org.springframework.kafka.core.KafkaTemplate;
-import org.springframework.kafka.support.SendResult;
 import org.springframework.stereotype.Service;
-import org.springframework.util.StringUtils;
-import org.springframework.util.concurrent.ListenableFuture;
-import org.springframework.util.concurrent.ListenableFutureCallback;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.File;
@@ -58,7 +41,9 @@ import java.nio.file.Paths;
 import java.time.LocalDateTime;
 import java.time.OffsetDateTime;
 import java.time.format.DateTimeFormatter;
-import java.util.*;
+import java.util.List;
+import java.util.Map;
+import java.util.UUID;
 
 import static it.pagopa.selfcare.mscore.constant.GenericError.GENERIC_ERROR;
 import static it.pagopa.selfcare.mscore.constant.GenericError.UNABLE_TO_DOWNLOAD_FILE;
@@ -68,35 +53,23 @@ import static it.pagopa.selfcare.mscore.core.util.PdfMapper.*;
 @Slf4j
 @Service
 public class ContractService {
-
-    static final String DESCRIPTION_TO_REPLACE_REGEX = " - COMUNE";
     private final PagoPaSignatureConfig pagoPaSignatureConfig;
     private final PadesSignService padesSignService;
     private final FileStorageConnector fileStorageConnector;
     private final CoreConfig coreConfig;
     private final SignatureService signatureService;
-    private final KafkaTemplate<String, String> kafkaTemplate;
-    private final KafkaPropertiesConfig kafkaPropertiesConfig;
     private final ObjectMapper mapper;
-    private final PartyRegistryProxyConnector partyRegistryProxyConnector;
-    private final InstitutionConnector institutionConnector;
 
     public ContractService(PagoPaSignatureConfig pagoPaSignatureConfig,
                            FileStorageConnector fileStorageConnector,
                            CoreConfig coreConfig,
                            Pkcs7HashSignService pkcs7HashSignService,
-                           SignatureService signatureService,
-                           KafkaTemplate<String, String> kafkaTemplate,
-                           KafkaPropertiesConfig kafkaPropertiesConfig,
-                           PartyRegistryProxyConnector partyRegistryProxyConnector,
-                           InstitutionConnector institutionConnector) {
+                           SignatureService signatureService) {
         this.pagoPaSignatureConfig = pagoPaSignatureConfig;
         this.padesSignService = new PadesSignServiceImpl(pkcs7HashSignService);
         this.fileStorageConnector = fileStorageConnector;
         this.coreConfig = coreConfig;
         this.signatureService = signatureService;
-        this.kafkaTemplate = kafkaTemplate;
-        this.kafkaPropertiesConfig = kafkaPropertiesConfig;
         this.mapper = new ObjectMapper();
         SimpleModule simpleModule = new SimpleModule();
         simpleModule.addSerializer(OffsetDateTime.class, new JsonSerializer<>() {
@@ -106,8 +79,6 @@ public class ContractService {
             }
         });
         mapper.registerModule(simpleModule);
-        this.partyRegistryProxyConnector = partyRegistryProxyConnector;
-        this.institutionConnector = institutionConnector;
     }
 
     public File createContractPDF(String contractTemplate, User validManager, List<User> users, Institution institution, OnboardingRequest request, List<InstitutionGeographicTaxonomies> geographicTaxonomies, InstitutionType institutionType) {
@@ -228,148 +199,6 @@ public class ContractService {
             return path.toFile();
         } catch (IOException e) {
             throw new InvalidRequestException(String.format(UNABLE_TO_DOWNLOAD_FILE.getMessage(), coreConfig.getLogoPath()), UNABLE_TO_DOWNLOAD_FILE.getCode());
-        }
-    }
-
-    public void sendDataLakeNotification(Institution institution, Token token, QueueEvent queueEvent) {
-        log.debug(LogUtils.CONFIDENTIAL_MARKER, "sendDataLakeNotification institution = {}, token = {}, queueEvent = {}", institution, token, queueEvent);
-        if (institution != null) {
-            NotificationToSend notification = toNotificationToSend(institution, token, queueEvent);
-            log.debug(LogUtils.CONFIDENTIAL_MARKER, "Notification to send to the data lake, notification: {}", notification);
-            try {
-                String msg = mapper.writeValueAsString(notification);
-                sendNotification(msg, token.getId());
-            } catch (JsonProcessingException e) {
-                log.warn("error during send dataLake notification for token {}", notification.getId());
-            }
-        }
-    }
-
-    public NotificationToSend toNotificationToSend(Institution institution, Token token, QueueEvent queueEvent) {
-        NotificationToSend notification = new NotificationToSend();
-        if (queueEvent.equals(QueueEvent.ADD)) {
-            // When Onboarding.complete event id is the onboarding id
-            notification.setId(token.getId());
-            notification.setState(RelationshipState.ACTIVE.toString());
-            // when onboarding complete last update is activated date
-            notification.setUpdatedAt(Optional.ofNullable(token.getActivatedAt()).orElse(token.getCreatedAt()));
-        } else {
-            // New id
-            notification.setId(UUID.randomUUID().toString());
-            notification.setState(token.getStatus() == RelationshipState.DELETED ? "CLOSED" : token.getStatus().toString());
-            // when update last update is updated date
-            notification.setUpdatedAt(Optional.ofNullable(token.getUpdatedAt()).orElse(token.getCreatedAt()));
-            if (token.getStatus().equals(RelationshipState.DELETED)) {
-                // Queue.ClosedAt: if token.deleted show closedAt
-                notification.setClosedAt(Optional.ofNullable(token.getDeletedAt()).orElse(token.getUpdatedAt()));
-                notification.setUpdatedAt(Optional.ofNullable(token.getDeletedAt()).orElse(token.getUpdatedAt()));
-            } else {
-                // when update last update is updated date
-                notification.setUpdatedAt(Optional.ofNullable(token.getUpdatedAt()).orElse(token.getCreatedAt()));
-            }
-        }
-        // ADD or UPDATE msg event
-        notification.setNotificationType(queueEvent);
-        return toNotificationToSend(notification, institution, token);
-    }
-
-    public NotificationToSend toNotificationToSend(NotificationToSend notification, Institution institution, Token token) {
-        notification.setInternalIstitutionID(institution.getId());
-        notification.setProduct(token.getProductId());
-        notification.setFilePath(token.getContractSigned());
-        notification.setOnboardingTokenId(token.getId());
-        // Queue.CreatedAt: onboarding complete date
-        notification.setCreatedAt(Optional.ofNullable(token.getActivatedAt()).orElse(token.getCreatedAt()));
-
-        // ADD or UPDATE msg event
-        notification.setFileName(token.getContractSigned() == null ? "" : Paths.get(token.getContractSigned()).getFileName().toString());
-        notification.setContentType(token.getContentType() == null ? "" : token.getContentType());
-
-        if (token.getProductId() != null && institution.getOnboarding() != null) {
-            Onboarding onboarding = institution.getOnboarding().stream()
-                    .filter(o -> token.getProductId().equalsIgnoreCase(o.getProductId()))
-                    .findFirst().orElseThrow(() -> new InvalidRequestException(String.format("Product %s not found", token.getProductId()), "0000"));
-            notification.setPricingPlan(onboarding.getPricingPlan());
-            notification.setBilling(onboarding.getBilling() != null ? onboarding.getBilling() : institution.getBilling());
-            notification.setInstitution(toInstitutionToNotify(institution));
-        }
-
-        return notification;
-    }
-
-    private InstitutionToNotify toInstitutionToNotify(Institution institution) {
-        InstitutionToNotify toNotify = new InstitutionToNotify();
-        toNotify.setInstitutionType(institution.getInstitutionType());
-        toNotify.setDescription(institution.getDescription());
-        toNotify.setDigitalAddress(institution.getDigitalAddress() == null? coreConfig.getInstitutionAlternativeEmail(): institution.getDigitalAddress());
-        toNotify.setAddress(institution.getAddress());
-        toNotify.setTaxCode(institution.getTaxCode());
-        toNotify.setOrigin(institution.getOrigin());
-        toNotify.setOriginId(institution.getOriginId());
-        toNotify.setZipCode(institution.getZipCode());
-        toNotify.setPaymentServiceProvider(institution.getPaymentServiceProvider());
-        if (institution.getSubunitType() != null && !institution.getSubunitType().equals("EC")) {
-            try {
-                InstitutionPaSubunitType.valueOf(institution.getSubunitType());
-                toNotify.setSubUnitType(institution.getSubunitType());
-                toNotify.setSubUnitCode(institution.getSubunitCode());
-            } catch (IllegalArgumentException ignored) {
-            }
-        }
-        RootParent rootParent = new RootParent();
-        rootParent.setDescription(institution.getParentDescription());
-        if (StringUtils.hasText(institution.getRootParentId())) {
-            rootParent.setId(institution.getRootParentId());
-            Institution rootParentInstitution = institutionConnector.findById(institution.getRootParentId());
-            rootParent.setOriginId(Objects.nonNull(rootParentInstitution) ? rootParentInstitution.getOriginId() : null);
-            toNotify.setRootParent(rootParent);
-        }
-
-        if (institution.getAttributes() != null && institution.getAttributes().size() > 0) {
-            toNotify.setCategory(institution.getAttributes().get(0).getCode());
-        }
-        if (institution.getCity() == null) {
-            setInstitutionLocation(toNotify, institution);
-        } else {
-            toNotify.setCounty(institution.getCounty());
-            toNotify.setCountry(institution.getCountry());
-            toNotify.setIstatCode(institution.getIstatCode());
-            toNotify.setCity(institution.getCity().replace(DESCRIPTION_TO_REPLACE_REGEX, ""));
-        }
-        return toNotify;
-    }
-
-    private void sendNotification(String message, String tokenId) {
-        ListenableFuture<SendResult<String, String>> future =
-                kafkaTemplate.send(kafkaPropertiesConfig.getDatalakeContractsTopic(), message);
-
-        future.addCallback(new ListenableFutureCallback<>() {
-
-            @Override
-            public void onSuccess(SendResult<String, String> result) {
-                log.info("sent dataLake notification for token : {}", tokenId);
-            }
-
-            @Override
-            public void onFailure(Throwable ex) {
-                log.warn("error during send dataLake notification for token {}: {} ", tokenId, ex.getMessage(), ex);
-            }
-        });
-
-    }
-
-    private void setInstitutionLocation(InstitutionToNotify toNotify, Institution institution) {
-        try {
-            InstitutionProxyInfo institutionProxyInfo = partyRegistryProxyConnector.getInstitutionById(institution.getExternalId());
-            toNotify.setIstatCode(institutionProxyInfo.getIstatCode());
-            toNotify.setCategory(institutionProxyInfo.getCategory());
-            GeographicTaxonomies geographicTaxonomies = partyRegistryProxyConnector.getExtByCode(toNotify.getIstatCode());
-            toNotify.setCounty(geographicTaxonomies.getProvinceAbbreviation());
-            toNotify.setCountry(geographicTaxonomies.getCountryAbbreviation());
-            toNotify.setCity(geographicTaxonomies.getDescription().replace(DESCRIPTION_TO_REPLACE_REGEX, ""));
-        } catch (MsCoreException | ResourceNotFoundException e) {
-            log.warn("Error while searching institution {} on IPA, {} ", institution.getExternalId(), e.getMessage());
-            toNotify.setIstatCode(null);
         }
     }
