+++ core/src/test/java/it/pagopa/selfcare/mscore/core/OnboardingServiceImplTest.java
@@ -67,6 +67,9 @@ class OnboardingServiceImplTest {
     private ContractService contractService;
 
     @Mock
+    private ContractEventNotificationService contractEventNotificationService;
+
+    @Mock
     private UserService userService;
     @Mock
     private MailNotificationService emailService;
