+++ core/src/test/java/it/pagopa/selfcare/mscore/core/TokenServiceImplTest.java
@@ -42,7 +42,7 @@ import static org.mockito.Mockito.when;
 @ExtendWith(MockitoExtension.class)
 class TokenServiceImplTest {
     @Mock
-    private ContractService contractService;
+    private ContractEventNotificationService contractService;
     @Mock
     private UserService userService;
